Fix construction of rulechain visitor
Ctor needs to be public, because RuleChain
builds them from a class
