+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/BaseLanguageModule.java
@@ -11,10 +11,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import net.sourceforge.pmd.Rule;
-import net.sourceforge.pmd.RuleContext;
-import net.sourceforge.pmd.lang.ast.Node;
-import net.sourceforge.pmd.lang.rule.AbstractRuleChainVisitor;
+import net.sourceforge.pmd.lang.internal.DefaultRulechainVisitor;
 import net.sourceforge.pmd.util.CollectionUtil;
 
 /**
@@ -202,28 +199,5 @@ public abstract class BaseLanguageModule implements Language {
         return getName().compareTo(o.getName());
     }
 
-
-
-    private static class DefaultRulechainVisitor extends AbstractRuleChainVisitor {
-
-        @Override
-        protected void visit(Rule rule, Node node, RuleContext ctx) {
-            rule.apply(Collections.singletonList(node), ctx);
-        }
-
-        @Override
-        protected void indexNodes(List<Node> nodes, RuleContext ctx) {
-            for (Node node : nodes) {
-                indexNodeRec(node);
-            }
-        }
-
-        protected void indexNodeRec(Node top) {
-            indexNode(top);
-            for (Node child : top.children()) {
-                indexNodeRec(child);
-            }
-        }
-    }
 
 }
